feat : 책 상세 정보 화면 만들기
책 상세 페이지를 컴포넌트 단위로 쪼개봄
