+++ src/components/book/book-image.scss
@@ -1,0 +1,4 @@
+img {
+  width: 400px;
+  margin-left: 50px;
+}
