+++ src/domain/book/Book.js
@@ -1,0 +1,11 @@
+class Book {
+  constructor(id, title, thumbnail, categories, bookRental) {
+    this.id = id;
+    this.title = title;
+    this.thumbnail = thumbnail;
+    this.categories = categories;
+    this.bookRental = bookRental;
+  }
+}
+
+export default Book;
